requirements.txt
numpy
pandas
matplotlib
librosa
keras
tensorflow
scikit-learn
soundfile
tqdm
pyyaml
